refactor: move logging_util module to training subpackage
